requirements.txt
#ipykernel==7.1.0
numpy == 2.4.0
matplotlib==3.10.0
pandas == 3.0.0
torchinfo == 1.8.0